producer/requirements.txt
kafka-python==2.0.2
faker==22.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
